Changed file names and reduced tests
